% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_index.r
\name{rand_index}
\alias{rand_index}
\title{Computes Rand index.}
\usage{
rand_index(a, b)
}
\arguments{
\item{a}{particion 1 as vector}

\item{b}{partition 2 as vector}
}
\value{
Rand index that ranges from 0 to 1
}
\description{
Computes Rand index.
}
